+++ packages/serializers/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tango-ts/serializers
 
+## 0.3.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @tango-ts/orm@0.3.0
+  - @tango-ts/core-types@0.3.0
+
 ## 0.2.2
 
 ### Patch Changes