+++ packages/views/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tango-ts/views
 
+## 0.3.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @tango-ts/orm@0.3.0
+  - @tango-ts/auth@0.3.0
+  - @tango-ts/core-types@0.3.0
+  - @tango-ts/http@0.3.0
+  - @tango-ts/router@0.3.0
+  - @tango-ts/serializers@0.3.0
+
 ## 0.2.2
 
 ### Patch Changes